+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix pointer reading for large address aware 32-bit executables.
+
 ## [0.8.2] - 2026-08-07
 
 ### Changed